chore(api): separate profile route to own module
